Remove window.location by fixing paypal error
We have to unmount the paypal button before change pages or else error
occur. Fix by putting in CartView, a function, handlePaypalHide(), and
state, displayPaypalButton, to pass to children Cart commponents and
their Checkout Component children. On successful payment, call the
parent function which changes the state which causes a render of the
parent and its children.